Combined_Backend/requirements.txt
# ──────────────────────────────────────────────
# FluentFusion AI — Unified Requirements
# Platform API + PULSE ML Engine
# ──────────────────────────────────────────────

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Config & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Async / Cache
redis==5.0.1
celery==5.3.4
httpx==0.25.2

# Payments
stripe==7.5.0

# Email
sendgrid==6.10.0

# File Storage
boto3==1.34.0

# ── PULSE ML Engine
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# Dev & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
isort==5.13.2
flake8==6.1.0
